add region fit api
